sửa lại security và nâng cấp json token, hàm mới nhận jwt và xác thực

+++ DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/Category.java
@@ -8,7 +8,7 @@ import jakarta.persistence.*;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer ID;  // ID tự tăng
+    private Long ID;  // ID tự tăng
 
     private String Name;
     private Boolean Status;
@@ -24,11 +24,11 @@ public class Category {
     }
 
     // Getters and Setters
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
